First Round Boggle Solver
User enters a boggle board.
To solve, it filters words out of local file word list(needs to be
replaced due to many missing words) to create a list of words only
containing letters on the board and of a valid length. Then iterates
through word list to check if each word can be made on the board.
Returns final list of words that can be made on the board to the user.
Bulk of solution can be found in app\Boggle.php